pyproject.toml
[tool.poetry]
name = "marvin-assistant"
version = "0.1.0"
description = ""
authors = ["Marcus Vinicus Braga <mvbraga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.5.2"
pyaudio = "^0.2.14"
pydub = "^0.25.1"
lameenc = "^1.6.3"
python-dotenv = "^1.0.0"
openai = "0.28.1"
langchain = "^0.0.340"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
